refactor: [#1195] extract authentication::Service
